requirements.txt
pandas
numpy
jupyter
papermill
nbconvert
nbformat
plotly
flask
gunicorn
google-cloud-storage